+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### 2.11.3
 - [30680 - Faixa com saldo do cashback na jornada do APP](https://dev.azure.com/ARECO/AReCO/_workitems/edit/30680)
-
 ---
+
+- [30688 - Inclusão do desconto do prime no valor final dos produtos na sacola do app](https://dev.azure.com/areco/AReCO/_workitems/edit/30688)
 
 ### 2.11.2
 - [31626 Tracking de eventos Algolia](https://dev.azure.com/ARECO/AReCO/_workitems/edit/31626)